Add ScreenTitleIcon class, used for displaying custom specific icons on overlays.
