+++ osu.Game/Overlays/Changelog/ChangelogHeader.cs
@@ -7,13 +7,11 @@ using osu.Framework.Allocation;
 using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
-using osu.Framework.Graphics.Shapes;
 using osu.Framework.Graphics.Sprites;
 using osu.Framework.Graphics.Textures;
 using osu.Game.Graphics;
 using osu.Game.Graphics.UserInterface;
 using osu.Game.Online.API.Requests.Responses;
-using osuTK;
 
 namespace osu.Game.Overlays.Changelog
 {
@@ -123,48 +121,7 @@ namespace osu.Game.Overlays.Changelog
                 AccentColour = colours.Violet;
             }
 
-            protected override Drawable CreateIcon() => new ChangelogIcon();
-
-            internal class ChangelogIcon : CompositeDrawable
-            {
-                private const float circle_allowance = 0.8f;
-
-                [BackgroundDependencyLoader]
-                private void load(TextureStore textures, OsuColour colours)
-                {
-                    Size = new Vector2(ICON_SIZE / circle_allowance);
-
-                    InternalChildren = new Drawable[]
-                    {
-                        new CircularContainer
-                        {
-                            Masking = true,
-                            BorderColour = colours.Violet,
-                            BorderThickness = 3,
-                            MaskingSmoothness = 1,
-                            RelativeSizeAxes = Axes.Both,
-                            Children = new Drawable[]
-                            {
-                                new Sprite
-                                {
-                                    RelativeSizeAxes = Axes.Both,
-                                    Texture = textures.Get(@"Icons/changelog"),
-                                    Size = new Vector2(circle_allowance),
-                                    Anchor = Anchor.Centre,
-                                    Origin = Anchor.Centre,
-                                },
-                                new Box
-                                {
-                                    RelativeSizeAxes = Axes.Both,
-                                    Colour = colours.Violet,
-                                    Alpha = 0,
-                                    AlwaysPresent = true,
-                                },
-                            }
-                        },
-                    };
-                }
-            }
+            protected override Drawable CreateIcon() => new ScreenTitleIcon(@"Icons/changelog");
         }
     }
 }
